requirements.txt
Django==4.1.7
openai==0.26.5
uvicorn~=0.20.0
psycopg2-binary~=2.9.5
python-dotenv~=0.21.1
dj-database-url~=1.2.0
djangorestframework~=3.14.0
tiktoken~=0.2.0
djangorestframework-simplejwt~=5.2.2
mysqlclient~=2.1.1